TransitionManifolds.jl

TransitionManifolds.jl

Compute Distances

Compute Embedding

Preprocessing

Helper functions

Compute Distances

TransitionManifolds.TransitionDistanceProblem — Type

TransitionDistanceProblem{T,W,L}(data; weights)

Struct for holding the data that is used to compute transition distances, see compute_distances.

The data contains burst simulation samples for n_anchor anchor points, and can be in one of two layouts L.

In the L=Contiguous layout, the data is an Array{T,3}, i.e., the number of samples n_samples is equal for each anchor, and the data has the shape (d, n_samples, n_anchors).

In the L=Jagged layout, the data is an Vector{Array{T,2}}, i.e., each anchor may have a different number of samples, it is length(data) = n_anchors, and size(data[i]) = (d, n_samples_i).

Optionally, weights can be specified that give each sample in data a weight. Thus, the layout and shape of weights has to match the data: In the L=Contiguous layout, the weights are an Array{W,2} of shape (n_samples, n_anchors), and in the L=Jagged layout, the weights are a Vector{Vector{W}} of length n_anchors.

The type of data points is T<:Real, and the type of the weights is W<:Real if weights were provided and W=Nothing otherwise.

TransitionManifolds.AbstractDataLayout — Type

AbstractDataLayout

Abstract type for the layout of data.

See also TransitionDistanceProblem and compute_distances.

TransitionManifolds.Contiguous — Type

Contiguous <: AbstractDataLayout

Contiguous data layout Array{T, 3}.

See also TransitionDistanceProblem and compute_distances.

TransitionManifolds.Jagged — Type

Jagged <: AbstractDataLayout

Jagged data layout Vector{Array{T,2}}.

See also TransitionDistanceProblem and compute_distances.

TransitionManifolds.layout — Function

layout(prob::TransitionDistanceProblem)

Layout of prob; either Contiguous or Jagged.

TransitionManifolds.compute_distances — Function

compute_distances(prob, alg::KernelDStatMMD; kwargs...) -> TransitionDistanceResult

The KernelDStatMMD algorithm works with Contiguous and Jagged layout. Weighted samples are not supported. The res.info dictionary contains

res.info["elapsed"]: the elapsed time

compute_distances(prob, alg::KernelVStatMMD; kwargs...) -> TransitionDistanceResult

The KernelVStatMMD algorithm works with Contiguous and Jagged layout. Weighted samples are not supported. The res.info dictionary contains

res.info["elapsed"]: the elapsed time

compute_distances(prob::TransitionDistanceProblem, alg::AbstractTransitionDistanceAlgorithm; progress=false) -> TransitionDistanceResult

Compute pairwise distances of transition density functions, returning a TransitionDistanceResult object res.

The distance between transition densities $p_x$ and $p_y$ for each pair of anchor points $x$ and $y$ is estimated using the specified algorithm alg, using the data in the given TransitionDistanceProblem prob.

The result res contains the pairwise distance matrix res.distances and the res.info dictionary, which is used to store further information, see the documentation for each specific algorithm.

compute_distances(data, alg::AbstractTransitionDistanceAlgorithm; weights=nothing, progress=false) -> TransitionDistanceResult

The data, and optionally weights, can be provided directly instead of a TransitionDistanceProblem.

The data contains burst simulation samples for n_anchor anchor points, and can be in one of two layouts.

In the Contiguous layout, the data is an Array{T,3}, i.e., the number of samples n_samples is equal for each anchor, and the data has the shape (d, n_samples, n_anchors).

In the Jagged layout, the data is an Vector{Array{T,2}}, i.e., each anchor may have a different number of samples, it is length(data) = n_anchors, and size(data[i]) = (d, n_samples_i).

The weights give each sample in data a weight. Thus, the layout and shape of weights has to match the data: In the L=Contiguous layout, the weights are an Array{W,2} of shape (n_samples, n_anchors), and in the L=Jagged layout, the weights are a Vector{Vector{W}} of length n_anchors.

compute_distances(pres::PreprocessResult, alg::AbstractTransitionDistanceAlgorithm; kwargs...) -> TransitionDistanceResult

A PreprocessResult pres can be provided directly instead of a TransitionDistanceProblem.

compute_distances(prob, alg::GaussianDStatMMD; kwargs...) -> TransitionDistanceResult

The GaussianDStatMMD algorithm works with Contiguous and Jagged layout. Weighted samples are not supported. The res.info dictionary contains

res.info["elapsed"]: the elapsed time

res.info["bandwidth"]: the used bandwidth

TransitionManifolds.TransitionDistanceResult — Type

TransitionDistanceResult{T<:Real}

Struct for storing the result of the transition distances computation, see compute_distances.

Distance Algorithms

TransitionManifolds.AbstractTransitionDistanceAlgorithm — Type

AbstractTransitionDistanceAlgorithm

Abstract type for algorithms for computing the distance between transition density functions.

See also compute_distances.

TransitionManifolds.GaussianDStatMMD — Type

GaussianDStatMMD(bandwidth=nothing) <: AbstractTransitionDistanceAlgorithm

Struct for using the maximum mean discrepancy (MMD) with a Gaussian kernel and D-Statistic estimation to compute the transition density distances.

The MMD between random variables $X$ and $Y$ is given by

\[E[k(X, X')] + E[k(Y, Y')] - 2 E[k(X, Y)].\]

Here, $k$ is a Gaussian kernel: $k(x, y) := exp(-||x - y||^2 / σ^2)$ where $σ$ is called the bandwidth. The D-Statistic is used to estimate the above expected values from samples. The computational cost is linear in the number of samples. For a more accurate estimator that scales quadratically in the number of samples, see GaussianVStatMMD.

The bandwidth is either a number > 0 or nothing, in which case a reasonable bandwidth is chosen automatically based on the samples.

TransitionManifolds.GaussianVStatMMD — Type

GaussianVStatMMD(...) <: AbstractTransitionDistanceAlgorithm

Not implemented yet!

TransitionManifolds.KernelDStatMMD — Type

KernelDStatMMD(kernel::Kernel) <: AbstractTransitionDistanceAlgorithm

Struct for using the maximum mean discrepancy (MMD) with any kernel from KernelFunctions and D-Statistic estimation to compute the transition density distances.

The MMD between random variables $X$ and $Y$ is given by

\[E[k(X, X')] + E[k(Y, Y')] - 2 E[k(X, Y)],\]

where $k$ is the kernel.

The D-Statistic is used to estimate the above expected values from samples. The computational cost is linear in the number of samples. For a more accurate estimator that scales quadratically in the number of samples, see KernelVStatMMD.

TransitionManifolds.KernelVStatMMD — Type

KernelVStatMMD(kernel::Kernel) <: AbstractTransitionDistanceAlgorithm

Struct for using the maximum mean discrepancy (MMD) with any kernel from KernelFunctions and V-Statistic estimation to compute the transition density distances.

The MMD between random variables $X$ and $Y$ is given by

\[E[k(X, X')] + E[k(Y, Y')] - 2 E[k(X, Y)],\]

where $k$ is the kernel.

The V-Statistic is used to estimate the above expected values from samples. The computational cost is quadratic in the number of samples. For a less costly and less accurate estimator see KernelDStatMMD.

Compute Embedding

TransitionManifolds.compute_embedding — Function

compute_embedding(distances::AbstractMatrix{<:Real}, alg::AbstractEmbeddingAlgorithm; n_coordinates, progress=false) -> EmbeddingResult

Compute an embedding from distances, returning a EmbeddingResult object res.

The distances should be a (n_anchors, n_anchors) symmetric distance matrix with 0 on the diagonal.

The desired dimension of the embedding can be specified using n_coordinates, but it is not guaranteed that the alg is able to provide exactly this number. By default, all computed coordinates are returned.

The result res contains the res.coordinates of the embedding in the shape (n_anchors, n_coordinates), and the res.info dictionary, which is used to store further information, see the documentation for each specific algorithm.

compute_embedding(dres::TransitionDistanceResult, alg::AbstractEmbeddingAlgorithm; kwargs...) -> EmbeddingResult

A TransitionDistanceResult dres can be provided instead of a distance matrix distances.

compute_embedding(distances, alg::DiffusionMaps; kwargs...) -> EmbeddingResult

When using the DiffusionMaps algorithm, the res.info dictionary contains

res.info["elapsed"]: the elapsed time

res.info["bandwidth"]: the used bandwidth

res.info["eigvals"]: the eigenvalues of the diffusion matrix

res.info["eigvecs"]: the eigenvectors of the diffusion matrix

res.info["dimension_estimate"]: estimate of the manifold dimension

TransitionManifolds.EmbeddingResult — Type

EmbeddingResult{T<:Real}

Struct for storing the result of the embedding computation, see compute_embedding.

Embedding Algorithms

TransitionManifolds.AbstractEmbeddingAlgorithm — Type

AbstractEmbeddingAlgorithm

Abstract type for algorithms for computing low-dimensional embeddings.

See also compute_embedding.

TransitionManifolds.DiffusionMaps — Type

DiffusionMaps(bandwidth=nothing, alpha=0.5) <: AbstractEmbeddingAlgorithm

Struct for using diffusion maps to compute an embedding.

In the diffusion maps algorithm, the similarity matrix $L$ is computed from the pairwise distance matrix $Δ$ via

\[L_{ij} := exp(-Δ_{i,j}^2 / σ^2)\]

where $σ$ is called the bandwidth. Then $L$ is normalized according to the normalization parameter $α$, i.e.,

\[D^{-α} L D^{-α}\]

where $D$ is the diagonal matrix containing the row-sums of $L$. The resulting matrix is then row-normalized to obtain the stochastic diffusion matrix $M$. Finally, the coordinates of the embedding are given as the dominant eigenvectors of $M$, multiplied by the associated eigenvalues.

The bandwidth is either a number > 0 or nothing, in which case a reasonable bandwidth is chosen automatically. Typical choices for the normalization parameter alpha are

α = 0: Graph Laplacian normalization

α = 0.5: Fokker-Planck normalization

α = 1: Laplace-Beltrami normalization.

Preprocessing

TransitionManifolds.preprocess — Function

preprocess(data::Array{Graph, 2}, mode=:adjacency_matrix) -> PreprocessResult

Convert an array of graphs of shape (n_samples, n_anchors) to an array of numbers by replacing each graph with a vector representation.

Supports the following modes for the conversion:

mode=:adjacency_matrix: replace each graph with its flattened adjacency matrix. All graphs must have the same number of nodes.

preprocess(data) -> PreprocessResult

Preprocess data so that it can be used in compute_distances, returning a PreprocessResult object res.

The result res contains a TransitionDistanceProblem at res.prob, and the res.info dictionary, which is used to store further information.

See the methods below associated to different data types.

preprocess(data::Trajectories; anchors, dist=Euclidean(), max_dist::Real, min_samples::Int, max_samples::Int) -> PreprocessResult

Obtain approximate burst simulation data from Trajectories.

The anchors should be provided as a (d, n_anchors) shaped Matrix. Then, for each anchor $a$ this function finds the trajectory points $x_i$ closer than max_dist in the metric dist (see Distances.jl), and adds the successors $x_{i+1}$ to the samples for $a$. Only the max_samples closest trajectory points are considered for each anchor (default: max_samples = ∞).

If no anchors are provided, it uses random samples from the trajectories. If no max_dist is provided, it uses half the average jump distance from the trajectories.

All anchors that by the end have less than min_samples samples are removed (default: min_samples=1).

The res.info dictionary contains

res.info["anchors"]: the final set of anchors

res.info["max_dist"]: the used max_dist

TransitionManifolds.PreprocessResult — Type

PreprocessResult

Struct for storing the result of preprocess.

TransitionManifolds.Trajectories — Type

Trajectories{T<:Real}(data)

Struct for storing one or multiple trajectories trajs of data type T.

The data is either a single trajectory in the form of a (d, n_points) shaped Matrix{T}, or multiple trajectories in the form of a Vector{Matrix{T}} where each trajectory has the same dimension d.

The length(trajs) returns the total number of points across all trajectories. Also supports iteration over all trajectory points, e.g., for point in trajs.

Indexing at i returns a view of the i-th point, 1 <= i <= length(trajs). (Example: if the first trajectory has 3 points, trajs[4] returns the first point of the second trajectory.)

Helper functions

TransitionManifolds.compute_transition_manifold — Function

compute_transition_manifold(prob::TransitionDistanceProblem, distance_alg::AbstractTransitionDistanceAlgorithm, embedding_alg::AbstractEmbeddingAlgorithm; n_coordinates, progress::Bool=false) -> Tuple{TransitionDistanceResult,EmbeddingResult}

Compute an embedding using the embedding_alg from the transition distances that are calculated using the distance_alg.

This is an auxiliary function that simply first calls compute_distances and then compute_embedding.

compute_transition_manifold(data, distance_alg::AbstractTransitionDistanceAlgorithm, embedding_alg::AbstractEmbeddingAlgorithm; weights=nothing, kwargs...) -> Tuple{TransitionDistanceResult,EmbeddingResult}

The data, and optionally weights, can be provided directly instead of a TransitionDistanceProblem, see also compute_distances.

compute_transition_manifold(pres::PreprocessResult, distance_alg::AbstractTransitionDistanceAlgorithm, embedding_alg::AbstractEmbeddingAlgorithm; kwargs...) -> Tuple{TransitionDistanceResult,EmbeddingResult}

A PreprocessResult pres can be provided directly instead of a TransitionDistanceProblem.

TransitionManifolds.cat_anchors — Function

cat_anchors(probs::TransitionDistanceProblem{T,W,Contiguous}...) -> TransitionDistanceProblem{T,W,Contiguous}

Concatenate multiple TransitionDistanceProblem of matching dimensions along the anchor dimension.

Only for Contiguous layout. For Jagged layout, see append_anchors!.

TransitionManifolds.append_anchors! — Function

append_anchors!(prob::TransitionDistanceProblem{T,W,Jagged}, probs::TransitionDistanceProblem{T,W,Jagged}...)

Append multiple TransitionDistanceProblem of matching dimensions onto prob, increasing the number of anchors.

Only for Jagged layout. For Contiguous layout, see cat_anchors.

TransitionManifolds.normalize_cloud — Function

normalize_cloud(X::AbstractMatrix{<:Real}; quant=0.99) -> Matrix

Center, rotate, and scale a point-cloud X of shape (n_points, n_coordinates) so that inliers (below the quant-th radius quantile) fit inside the unit cube.

Returns the normalized point-cloud.